requirements.txt
##Librerias##
#Python
pandas==1.4.2
geopy==2.2.0
Unidecode==1.3.4
#Django
Django==3.2.2
requests==2.25.1
djangorestframework==3.12.4
django-filter==2.4.0
django-rest-auth==0.9.5
Pillow==7.1.2
# Drivers para trabajar con Postgres en linux:
psycopg2-binary>=2.8
